refactor: split HandleFatalCrash.java into two different use cases: HandleOutOfMemoryError & HandleUnsupportedException
- implement unit tests
- retrieve and execute use cases in Main via DefaultConfig
